Makes the HomeContentToggle sticky and visible in more places
