+++ src/components/Filters/FeedFilters/BountyFeedFilters.tsx
@@ -1,0 +1,15 @@
+import { Group, GroupProps } from '@mantine/core';
+import { BountyFiltersDropdown } from '~/components/Bounty/Infinite/BountyFiltersDropdown';
+import { SortFilter } from '../SortFilter';
+import { useFeedFiltersStyles } from './FeedFilters.styles';
+
+export function BountyFeedFilters({ ...groupProps }: GroupProps) {
+  const { classes } = useFeedFiltersStyles();
+
+  return (
+    <Group className={classes.filtersWrapper} spacing={8} noWrap {...groupProps}>
+      <SortFilter type="bounties" variant="button" />
+      <BountyFiltersDropdown />
+    </Group>
+  );
+}
