+++ src/components/Filters/FeedFilters/FeedFilters.styles.ts
@@ -1,0 +1,11 @@
+import { createStyles } from '@mantine/core';
+
+export const useFeedFiltersStyles = createStyles((theme) => ({
+  filtersWrapper: {
+    [theme.fn.smallerThan('sm')]: {
+      width: '100%',
+
+      '> *': { flexGrow: 1 },
+    },
+  },
+}));
