+++ src/pages/articles/index.tsx
@@ -1,19 +1,14 @@
-import { createStyles, Group, Stack, Title, useMantineTheme } from '@mantine/core';
+import { Stack, Title } from '@mantine/core';
 
 import { Announcements } from '~/components/Announcements/Announcements';
 import { useArticleQueryParams } from '~/components/Article/article.utils';
 import { ArticleCategoriesInfinite } from '~/components/Article/Categories/ArticleCategoriesInfinite';
 import { ArticleCategories } from '~/components/Article/Infinite/ArticleCategories';
-import { ArticleFiltersDropdown } from '~/components/Article/Infinite/ArticleFiltersDropdown';
 import { ArticlesInfinite } from '~/components/Article/Infinite/ArticlesInfinite';
-import { SortFilter, ViewToggle } from '~/components/Filters';
-import { FullHomeContentToggle } from '~/components/HomeContentToggle/FullHomeContentToggle';
-import { HomeContentToggle } from '~/components/HomeContentToggle/HomeContentToggle';
 import { MasonryContainer } from '~/components/MasonryColumns/MasonryContainer';
 import { MasonryProvider } from '~/components/MasonryColumns/MasonryProvider';
 import { Meta } from '~/components/Meta/Meta';
 import { env } from '~/env/client.mjs';
-import { useFeatureFlags } from '~/providers/FeatureFlagsProvider';
 import { useFiltersContext } from '~/providers/FiltersProvider';
 import { constants } from '~/server/common/constants';
 import { createServerSideProps } from '~/server/utils/server-side-helpers';
@@ -32,20 +27,7 @@ export const getServerSideProps = createServerSideProps({
   },
 });
 
-const useStyles = createStyles((theme) => ({
-  filtersWrapper: {
-    [containerQuery.smallerThan('sm')]: {
-      width: '100%',
-
-      '> *': { flexGrow: 1 },
-    },
-  },
-}));
-
 export default function ArticlesPage() {
-  const { classes } = useStyles();
-  const theme = useMantineTheme();
-  const features = useFeatureFlags();
   const storedView = useFiltersContext((state) => state.articles.view);
   const { query } = useArticleQueryParams();
 
@@ -64,31 +46,16 @@ export default function ArticlesPage() {
         maxSingleColumnWidth={450}
       >
         <MasonryContainer fluid>
-          {query.favorites && <Title>Your Bookmarked Articles</Title>}
           <Stack spacing="xs">
             <Announcements
-              sx={(theme) => ({
+              sx={() => ({
                 marginBottom: -35,
                 [containerQuery.smallerThan('md')]: {
                   marginBottom: -5,
                 },
               })}
             />
-
-            <Group position="apart" spacing={8}>
-              {features.alternateHome ? <FullHomeContentToggle /> : <HomeContentToggle />}
-              <Group className={classes.filtersWrapper} spacing={8} noWrap>
-                <SortFilter type="articles" variant="button" />
-                <ArticleFiltersDropdown />
-                <ViewToggle
-                  type="articles"
-                  color="gray"
-                  radius="xl"
-                  size={36}
-                  variant={theme.colorScheme === 'dark' ? 'filled' : 'light'}
-                />
-              </Group>
-            </Group>
+            {query.favorites && <Title>Your Bookmarked Articles</Title>}
             {view === 'categories' ? (
               <ArticleCategoriesInfinite filters={query} />
             ) : (
