+++ src/pages/events/index.tsx
@@ -1,16 +1,13 @@
-import { createStyles, Stack, useMantineTheme, Center, Loader, Group } from '@mantine/core';
+import { Center, Loader, Stack } from '@mantine/core';
 
 import { Announcements } from '~/components/Announcements/Announcements';
 import { ArticleCard } from '~/components/Cards/ArticleCard';
-import { FullHomeContentToggle } from '~/components/HomeContentToggle/FullHomeContentToggle';
-import { HomeContentToggle } from '~/components/HomeContentToggle/HomeContentToggle';
 import { MasonryContainer } from '~/components/MasonryColumns/MasonryContainer';
 import { MasonryGrid } from '~/components/MasonryColumns/MasonryGrid';
 import { MasonryProvider } from '~/components/MasonryColumns/MasonryProvider';
 import { Meta } from '~/components/Meta/Meta';
 import { NoContent } from '~/components/NoContent/NoContent';
 import { env } from '~/env/client.mjs';
-import { useFeatureFlags } from '~/providers/FeatureFlagsProvider';
 import { constants } from '~/server/common/constants';
 import { createServerSideProps } from '~/server/utils/server-side-helpers';
 import { containerQuery } from '~/utils/mantine-css-helpers';
@@ -24,7 +21,6 @@ export const getServerSideProps = createServerSideProps({
 });
 
 export default function EventsPage() {
-  const features = useFeatureFlags();
   const { data, isLoading } = trpc.article.getEvents.useQuery(undefined, {
     trpc: { context: { skipBatch: true } },
   });
@@ -46,17 +42,13 @@ export default function EventsPage() {
         <MasonryContainer fluid>
           <Stack spacing="xs">
             <Announcements
-              sx={(theme) => ({
+              sx={() => ({
                 marginBottom: -35,
                 [containerQuery.smallerThan('md')]: {
                   marginBottom: -5,
                 },
               })}
             />
-            <Group position="apart" spacing={8}>
-              {features.alternateHome ? <FullHomeContentToggle /> : <HomeContentToggle />}
-            </Group>
-
             {isLoading ? (
               <Center p="xl">
                 <Loader size="xl" />
