+++ src/pages/models/index.tsx
@@ -1,27 +1,20 @@
-import { ActionIcon, Button, Group, Popover, Stack, Title, createStyles } from '@mantine/core';
-import { IconExclamationMark } from '@tabler/icons-react';
+import { Stack, Title } from '@mantine/core';
 import { Announcements } from '~/components/Announcements/Announcements';
 import { CategoryTags } from '~/components/CategoryTags/CategoryTags';
-import { SortFilter, ViewToggle } from '~/components/Filters';
-import { FullHomeContentToggle } from '~/components/HomeContentToggle/FullHomeContentToggle';
-import { HomeContentToggle } from '~/components/HomeContentToggle/HomeContentToggle';
 import { IsClient } from '~/components/IsClient/IsClient';
 import { MasonryContainer } from '~/components/MasonryColumns/MasonryContainer';
 import { MasonryProvider } from '~/components/MasonryColumns/MasonryProvider';
 import { Meta } from '~/components/Meta/Meta';
 import { ModelCategoriesInfinite } from '~/components/Model/Categories/ModelCategoriesInfinite';
-import { ModelFiltersDropdown } from '~/components/Model/Infinite/ModelFiltersDropdown';
 import { ModelsInfinite } from '~/components/Model/Infinite/ModelsInfinite';
 import { useModelQueryParams } from '~/components/Model/model.utils';
 import { env } from '~/env/client.mjs';
-import { hideMobile, showMobile } from '~/libs/sx-helpers';
-import { useFeatureFlags } from '~/providers/FeatureFlagsProvider';
 import { useFiltersContext } from '~/providers/FiltersProvider';
 import { constants } from '~/server/common/constants';
 import { PeriodMode } from '~/server/schema/base.schema';
 import { createServerSideProps } from '~/server/utils/server-side-helpers';
-import { QS } from '~/utils/qs';
 import { containerQuery } from '~/utils/mantine-css-helpers';
+import { QS } from '~/utils/qs';
 
 export const getServerSideProps = createServerSideProps({
   useSession: true,
@@ -39,19 +32,7 @@ export const getServerSideProps = createServerSideProps({
   },
 });
 
-const useStyles = createStyles(() => ({
-  filtersWrapper: {
-    [containerQuery.smallerThan('sm')]: {
-      width: '100%',
-
-      '> *': { flexGrow: 1 },
-    },
-  },
-}));
-
 export default function ModelsPage() {
-  const features = useFeatureFlags();
-  const { classes, theme } = useStyles();
   const storedView = useFiltersContext((state) => state.models.view);
   const { set, view: queryView, ...queryFilters } = useModelQueryParams();
   const { username, favorites, hidden, query, collectionId } = queryFilters;
@@ -75,9 +56,6 @@ export default function ModelsPage() {
         maxSingleColumnWidth={450}
       >
         <MasonryContainer fluid>
-          {username && typeof username === 'string' && <Title>Models by {username}</Title>}
-          {favorites && <Title>Your Liked Models</Title>}
-          {hidden && <Title>Your Hidden Models</Title>}
           <Stack spacing="xs">
             <Announcements
               sx={() => ({
@@ -87,37 +65,9 @@ export default function ModelsPage() {
                 },
               })}
             />
-            <Group position="apart" spacing={8}>
-              {features.alternateHome ? <FullHomeContentToggle /> : <HomeContentToggle />}
-              <Group className={classes.filtersWrapper} spacing={4}>
-                {periodMode && (
-                  <Popover>
-                    <Popover.Target>
-                      <ActionIcon variant="filled" color="blue" radius="xl" size={36} mr={4}>
-                        <IconExclamationMark size={20} strokeWidth={3} />
-                      </ActionIcon>
-                    </Popover.Target>
-                    <Popover.Dropdown maw={300}>
-                      {`To ensure that you see all possible results, we've disable the period filter.`}
-                      <Button mt="xs" size="xs" fullWidth onClick={() => set({ query: undefined })}>
-                        Clear Search
-                      </Button>
-                    </Popover.Dropdown>
-                  </Popover>
-                )}
-                <SortFilter type="models" variant="button" />
-                <ModelFiltersDropdown />
-                {canToggleView && (
-                  <ViewToggle
-                    type="models"
-                    color="gray"
-                    radius="xl"
-                    size={36}
-                    variant={theme.colorScheme === 'dark' ? 'filled' : 'light'}
-                  />
-                )}
-              </Group>
-            </Group>
+            {username && typeof username === 'string' && <Title>Models by {username}</Title>}
+            {favorites && <Title>Your Liked Models</Title>}
+            {hidden && <Title>Your Hidden Models</Title>}
             <IsClient>
               {view === 'categories' ? (
                 <ModelCategoriesInfinite />
