+++ src/pages/videos/index.tsx
@@ -1,8 +1,5 @@
-import { createStyles, Group, Stack, Title } from '@mantine/core';
+import { Stack, Title } from '@mantine/core';
 import { Announcements } from '~/components/Announcements/Announcements';
-import { SortFilter } from '~/components/Filters';
-import { FullHomeContentToggle } from '~/components/HomeContentToggle/FullHomeContentToggle';
-import { HomeContentToggle } from '~/components/HomeContentToggle/HomeContentToggle';
 import { ImageCategories } from '~/components/Image/Filters/ImageCategories';
 import { useImageFilters } from '~/components/Image/image.utils';
 import ImagesInfinite from '~/components/Image/Infinite/ImagesInfinite';
@@ -10,25 +7,11 @@ import { IsClient } from '~/components/IsClient/IsClient';
 import { MasonryContainer } from '~/components/MasonryColumns/MasonryContainer';
 import { MasonryProvider } from '~/components/MasonryColumns/MasonryProvider';
 import { Meta } from '~/components/Meta/Meta';
-import { useFeatureFlags } from '~/providers/FeatureFlagsProvider';
 import { constants } from '~/server/common/constants';
 import { env } from '~/env/client.mjs';
-import { VideoFiltersDropdown } from '~/components/Image/Filters/VideoFiltersDropdown';
-
-const useStyles = createStyles((theme) => ({
-  filtersWrapper: {
-    [theme.fn.smallerThan('sm')]: {
-      width: '100%',
-
-      '> *': { flexGrow: 1 },
-    },
-  },
-}));
 
 export default function VideosPage() {
-  const features = useFeatureFlags();
   const { hidden, ...filters } = useImageFilters('videos');
-  const { classes } = useStyles();
 
   return (
     <>
@@ -53,13 +36,6 @@ export default function VideosPage() {
                 },
               })}
             />
-            <Group position="apart" spacing={8}>
-              {features.alternateHome ? <FullHomeContentToggle /> : <HomeContentToggle />}
-              <Group className={classes.filtersWrapper} spacing={8} noWrap>
-                <SortFilter type="videos" variant="button" />
-                <VideoFiltersDropdown />
-              </Group>
-            </Group>
             <IsClient>
               <ImageCategories />
               <ImagesInfinite filters={{ ...filters, types: ['video'] }} showEof />
