+++ src/server/services/feature-flags.service.ts
@@ -35,13 +35,7 @@ const featureFlags = createFeatureFlags({
     description: `We're improving our search experience! This is a beta feature, so please report any issues you find!`,
     availability: ['public'],
   },
-  alternateHome: {
-    toggleable: true,
-    default: true,
-    displayName: 'New Home Page',
-    description: `A new home page with a more modern design and more features. This is a beta feature, so please report any issues you find!`,
-    availability: ['public'],
-  },
+  alternateHome: ['public'],
   collections: ['public'],
   air: {
     toggleable: true,
